repo: more granular forces module
