+++ crates/buoy-physics/README.md
@@ -1,3 +1,5 @@
-# buoy-core
+# buoy-physics
 
-This is the core simulation engine for the Buoyancy Simulator.
+This is the core simulation engine for buoyancy simulator, buoy. It contains the
+core physics calculations for buoyancy, drag, weight, ideal gas law, and basic
+elasticity.
