--- crates/buoy-physics/src/objects/mod.rs
@@ -1,5 +1,0 @@
-pub mod balloon;
-pub mod environment;
-
-pub use balloon::Balloon;
-pub use environment::GroundPlane;
